Add .gitignore statements for Sphinx
The .gitignore file in the source folder is adapted to not to version
Sphinx related build files, in order to keep the repository clean. I am
not 100% sure, whether those are the correct files to ignore, but there
is hardly information to be found on the internet. The test Sphinx
folders still include those ignored files.