Add use case and explanations
Use case 2 was added as a more advanced use case compared to the first
one. It is still an abstract one, but contains branches, hierarchy,
concurrency and a variety of different data flows.
The supplementary files explain the file/folder structure for this use
case (File and folder structure.doc) as well as giving a detailed
instructions of what is happening in the example state D, including some
code.
Microsoft Office was used instead of LibreOffice, as the latter had
problems with opening/saving files and constantly crashed...