Update the flow chart
The flow chart of the state machine was updated in order to incorporate
the changes made to the class diagram. A state machine is no longer
existing, but a plain state and a state container, both with different
run methods. The start/exit methods are no longer needed to distinguish
different types of states, but allow user to define scripts being called
at the start/exit point of a container state.