Add initial structure and classes for MVC architecture
The package mvc contains the files for the MVC architecture, being the
mediating layer between the state-machine and the GUI. The package
consists of three subpackages for the three parts model, controller and
view.
A controller, model and view for the plain state class was created.
Using the main module in mvc, the system can be tested.
The state class was modified to include the Observable pattern. This is
needed in order for other classes to get informed about changes to
attributes.
The documentation was created and added to index.rst
Another package in the root folder called support is intended for helper
classes, which are needed by several other classes. For now, it contains
a log module, allowing a consistent logging output for all modules. This
should be used rather than printf commands.