Add selection detection
When the mouse is clicked, the elements beneath the mouse (states,
transitions, data flows) are detected. The three types of elements are
put in seperate lists. This allows for highlighting selected elements as
well as drag'n'drop actions.
In addition, the viewport is now automatically updated, when initially
loading the states.