Removes observables from modify methods
The modify methods shouldn't be observed directly. The actual change
happens in the transitions or data flows themselves. Therefore only the
setters in the data flows/transitions are now observed and changed in
the models of these connections are propagated upwards to the container
state and further. This has also the advantage, that the correct model is
passed to the observing methods.