Update models before starting to notify
Up till now, there were two methods inside the (container) state model
listening for changes to the property state. The first method notified other
lists and the parent about these changes, the second method updated the
models. However, it was not determined, in which order these two methods
were executed. This could sometimes lead to errors, when first the
notifications were triggered and then the models were updated, because
the changes were not yet visible to the controllers listening for the
notifications.
This is now fixed by only letting the first method listen for changes
and when being called, first call the second method to update the
models.