Refactor waypoint handling of new connections
The handling of the addition of waypoints to newly created connections
was vastly simplified. The initial complex behaviour was based on a
misconception. Waypoints are added to the meta data of the connection
model directly after the creation of the new connection. This can be
done, as the according model is created automatically and directly after
the core class creation, before the application continues its execution.