Fix bug of erroneous outcome deletion
When deleting an outcome, transitions connected to this outcome have to
be deleted. This was done, but the check whether a transition was
connected to the outcome only looked at the from/to_outcome id, and not
at the from/to_state id. As the outcome id is not unique in this
context, wrong transitions were deleted in some scenarios. This is fixed
now.
- Also compare state id when removing transition for deleted outcome
- Add remove_outcome_hook to only add additional logic in the container
  state and not having to copy all the code from the state base class
- Introduce update method in transition controller to keep model and
  tree store consistent