Add HTML and JS files for browser control + HTTP Server
This commit adds support for Browser based control of the statemachine and
provides basic debug support.
The port of the HTTP server can be specified in the server_config file.
All commands coming from the Browser are forwarded to the statemachine by the
server and vice versa (messages to the server forwarded to browser).