Update dialog libraries for key interaction
The "Show dialog" library now accepts a further parameter, defining a
key mapping from keys to buttons. Multiple keys can be mapped to a
single button. If a key mapped to a button is clicked, a click on that
button is emulated.
The other dialog libraries were updated to make use of this.
All dialog libraries have a description now.