Set static border width and improve state selection
The border width is now static for a given hierarchy level. It can be adjusted
by changing the parameters ROOT_WIDTH and WIDTH_HIERARCHY_FACTOR in constants.py

The state is now selected when clicking on its name. A second click selects the
name and highlights it with a slightly brighter background to make it easier to
see if the name or the state is selected.