Introduce check_child_validity in state class
The state class is responsible for checking the extended validity of
its children, e.g. to prevent duplicate names or ids. For this, the
check_child_validity method is introduced, checking for those issues.
There is only one method for all kinds of children (outcomes, ports),
the method itself checks for the type of child and chooses the
appropriate tests.
In addition:
- state outcomes controller:
  * adapt to new exception type
  * do not update when getting focus (led to exception and is not
    necessary, as all properties are observed)
- state class:
  * get rid of _used_outcome_ids, as the keys of the outcomes dict
    already hold this information
- outcome class:
  * get rid of function handler for name checking
  * use new check_child_validity method instead