Fix menu bar controller action callbacks
Action callbacks, which are used with idle_add (e.g. in tests)
cannot return True, as this would cause idle_add to call the
method over and over again. Therefore, global callback that want
to return True, in order to prevent the shortcut to be passed on,
must be wrapped. This is done in the menu bar controller.