Fix undesired behaviour of state machine tree
In some cases, the state machine tree controller changed the current
selection. If e.g. a state and a transition gets selected, the
controller changes the selection to only the state. This is not
desired, thus the responsible line is removed in this commit. No side
effects were observed and also the tests run successfully.

However, the commit should be reviewed @Rico!