New test utility to count warnings and errors
A new method in test_utils allows for counting warning and errors that
have been logged. It can therefore be used to verify that there was no
error/warning or to check for a certain number of expected errors/
warnings.
The method requires pytest_capturelog to be installed:
pip install --user pytest-capturelog
To use the method in your tests, add 'caplog' as argument of the test_*
method and pass it to assert_logger_warnings_and_errors as first
argument.