Use RectanglePointPort for all ports
Use the new point port class for each port (income, outcome, input, output,
scoped variable). The port_side_size method keeps the size of the port
updated. The scoped variable ports needs special treatment, as the size of
this type of port is depending on the name of the port. Thus the size of
the port is updated in the draw method, where the dimensions are determined.