Make meta data observable
A new property of state models a state machine models is signal
called meta_signal. Whenever the meta data changes, the changing
instance is supposed to emit a signal with a string explaining the
cause.
The change of meta data of a state is bubbled up to the state machine,
where a state machine wide signal signalizes the change of any meta
data.