Check for state machine before accessing it
_mark_state_machine_as_dirty checks whether the state machine id is
still existing before using it to set the mark_dirty flag.

I don't really know, why this is necessary, but this fixed problems
in the tests. It seems to be a problem with idle_add, as the models
are derived from ModelMT, using idle_add to call the listening
methods.