Improve propagation of meta_signal
The meta_signal, which is propagated to the root state and finally
 to the state machine retrieves the notification properties of the
state, who's meta data was changed. The previous implementation
used the notification properties of the root state.
This required the notification change to be prepended with a fixed
string, when propagating the signal from the root state to the
state machine. There, the prefix is removed again.