Gaphas: Correctly store/load meta data of ports
The meta data of ports is now stored in the meta data of the
corresponding port model and not the state model. It is
therefore also retrieved from the port model.
Meta data of elements not having a model (name, income) are
added to the state model, but at the end of ['gui']['editor_
gaphas'], instead of the beginning.