New base class for state elements of core
A new base class StateElement serves as base class for all state
elements (ports, connections), similar to StateElementModel inside
the MVC architecture. This allows to handle common code.
In addition, the reference to the parent state of these state
elements is now also a weak reference, to avoid circular references.