Allow input data port connection to output data port
Before this commit, an exception was raised when an input data port of a
HierarchyState was directly connected using a data flow to an output data
port. The code assumed that data flows from input data ports with target
state the container state are always connected to scoped variables.
This is now fixed.