Added resize functionality to un-docked windows.
After removing the decoration for un-docked windows in order to use the
Top Toolbar used in the Main Window, the resize functionality was lost.
It has now been restored by adding a Status Bar at the bottom of the
undocked_window glade file. The 'has_resize_grip' property of the Status
Bar is set to True and each un-docked window is resizable from its
bottom right corner.