ExecutionEngine: Notify condition on all events except pause
The conditional variable "execution_condition_variable" is now
notified for all status changes, except pause. In order to do so also
for pause @brun_sb will have a look at the overall execution logic.

This makes it possible, to use this condition to wait for the start,
step and stop of the execution.

The logic changing the status was now moved into the already existing
"set_execution_mode" method, which is observable.