Gaphas: Set dynamic minimum sizes
StateViews and NameViews now have a minimum size restriction, which is
depending on the size of the parent. The maximum ratios are defined by
constants. NameViews can e.g. only be 1/8th of the size of its parent.

When moving states around and to the edge of their parent, the size is
no longer changing.