Add libraries to shortcut folder list in file dialogs
On the left hand side of file dialogs, there is a list with shortcut
folders for quick access, which shows some standard entries and can be
managed by the user. However, it is also possible to add custom
directories programmatically. This is used to show all root paths of
all libraries. This allows fast access to open or save libraries.