LibraryManager: Remember skipped library roots
Instead of remembering only skipped libraries, now also library roots
(such as "generic") are remembered. This drastically reduces the
number of dialogs a user has to click when opening a state machine
with a missing library root.

Additionally, the skipped and replaced libraries are resetted when
the library manager is initialized/updated.