Do not change selection is recreate_state_editor
After a state type change, the recreate_state_editor reopens the
state editor with the newly created state model. This model was also
put in the selection. This is now no longer the case, as the method
raising the state type change signal already puts the state model
into the selection after the signalization and thus after all
controllers had the opportunity to react to the signal.

Especially the GraphicalEditor can now create the view for the new
model, before the selection is changed.