add rafcon.utils.resources
extends the functionality of pkg_resources to also search for resource files in a predefined list of paths
this is necessary if rafcon is installed via pypi
resources located in the devel space (i.e. the git repository) are used in the first place before files located in e.g. "~/.local/share/rafcon"