refactor selection of tree widget controller to to prepare it for the new scheme
Update also dostrings and naming as well as refactor and reduce code
that is no more necessary or now better by the changes.
Tests are not running with the new selection method in use thats way it
is not used already. That has to be investigated.
Tree view updates in case of inconstitency state machine selection prior.