feat: first integration of tree view tab-key control
Integrate the Tab-Key first for the semantic data editor.
Tab-key control is possible on one tree-level.

Reduce some methods and varaibles in tree-view and list-view controller.
Refactoring of add, remove and key and value-edit methods of semantic
data editor to reduce complexity and more direct access of data.

No more change of dict-values or -keys if not changed to reduce
modification history action-items and notifications.