Tests: Refactor and improve test_plugins_example
I started touching the test as it used a fixed sleep time to wait for
the GUI to come up, which often led to issues. I therefore replaced the
sleep by a loop a and poller using select. It is ensured that the test
fails if the GUI hasn't started within 10s. In addition, the output is
now better udnerstandable.