fix(gaphas): Expose event on destruction
If a model had been selected when the GUI was closed, that element was
drawn, as it was deselected by the Selection. This had led (under some
circumstances) to exception during the close operation.

This is now fixed by relieving the Selection in the ExtendedGTKView when
then root state is being destroyed.