fix(test_restore_session): adjust to be more robust and cover more cases
The old version was aspecting that a restored state machine with
different core hash would have a path.
But a state machine would not have one if not saved already.
It seems that a simple new state machine with only a root state could also have different
hashs before saving and after restoring what cause a traceback while generating a info print.
The test is still tolerating hash variations which is bad but currently a general issue of RAFCON.