fix(restore sessione): test and arespective module
The module needed a wait for gui to secure the generation of the
state machine model before selecting those. The test it self
is mostly adapted to fix timing and adding the wait_for_gui at the right place
and with the call_gui_call_back to do this in the right thread.