fix(test_interface): account for gtk patching behavior
As long as the gtk module keeps constant the gtk main thread will always have the same thread id!
But, if the module is patched as in "test_interface.py" the gtk thread will get another thread id!
Thus, if both threads are in used_gui_threads then we simply allow this case
in the __notify_observer__ method