feat/test(destruction): imporve and adapt the searched_class and graph generation
The prints are improved and the searched class is set at the beginning and
can also be searched explizit. Additionally all aobject that are in the target_objects list
will generate a graph of referrers and all be collected in a class specific user save path in
the test folder.