fix(tree view controller) destruction by workaround gtk and gtkmvc
The gtkmvc observer are not cleaning up there reference lists after being unregistered
so this is performed by name mangling into the observer code.
The gtk widgets are not recursively disconnect all of the sub-widgets and also do not
run the destroy method of those so the disconnect or unregister is performed by the
abstract class TreeViewController. Here disconnect is sometimes used because destroy
would cause additional problems, see commented line of code.