fix(type change and gaphas): fix problems and add again work arounds to get it functional
The goal was firstable to have a working version and debug the work arounds from this again.
The adapt_complex_action method of the gaphas editor is now not removing elements and
gets the parent_v handed by the signal after callbacks. Here refactoring is also needed
but this is maybe the best if taking into account the work arounds, too.
For now the gaphas editor removes affected vies before the type change actually happens.