fix/tests(group/ungroup): debug destruction of left over core and model elements
Improve tests to cover also the procedures of menu-bar and both group/ungroup test cases.
Therefore the models implement also the recursive prepare_destruction flag.
Allow test-menu-bar to skip refresh of libraries and state machines which is now a left over todo.
The testing utils is now freeing the exception info dicts that are hold by the pytest.caplog.records
to allow destruction check like if only the rafcon instance is running.
Print the comparison of error and warning logs by default for fast debugging of those counts.