fix(gaphas/destrution/mod-history): action signal handling, cleanup and add some work arounds
The work arounds are there to firstable to tolerate the better format of the action signal
handling. The destruction for change_root_state_type/change_state_type, group/ungroup and copy/cut/paste
are working. There are several left over prints (commented), logger messages and work arounds for debugging reasons.
Those are left in to let the system run and get a over all picture of the problems instate on hanging always
on the first occuring traceback or problem.

Over all tests there are 48 'View is missing*' and 83 '*is already existing in*' logger messages
which needs to be solved. Several of them are in the copy/paste/cut actions. Also the gtkmvc.observable
try-except block is catching several tracebacks which are commenly not printed and get created by
copy/paste/cut operations e.g. in test_copy_methods.py.