fix(refresh-all destruction) on core and model level
The gaphas views are not correctly destroyed and therefore
not raised in the test but printed while the check of still exsisting objects.
Gaphas add again the work around of cleaning the _core_view_map and
the _model_view_map dictionaries.
The modification history has had references on the sm and root state models
which are now released.