fix(list controller): simplify work around and destruction_signal observation issue
The observation of the destruction_signal seems to be an problem in case of the list controllers
and especially the outcomes controller.
The controller work around checks now for the core_element is None case in the models
and avoids to update those elements. Still not full fixed but the update of self.model respected list
elements are preserved to work again. This issue can be also fixed after the pull request is merged already.