chore(library model destruction): add more details on issue
The issue has to be solved before merging the pull-request.
It seems to maybe destroy consistency of models and can cause future problems.

Best a refactoring of the ungroup and group helper methods to avoid this effects
and removal of all prepare destruction calls in insert meta data and avoiding those
method at all.