chore(library state model): tolerate multiple calls of destruction by default
The TODO was removed and the reason found and now multiple calls are explicit
tolerated but logged. We tolerate multiple calls of state models in general implicit
but in the library state model we have to do this explicit to avoid problems.
The multiple calls maybe can be come an issue and should be observed in general.