fix/refactor(group/ungroup): by expected future models, cleanup, more clear handling, checks
The gui.helper.state_machine checks for unique parent.
The gui.helper.state hands the expected models in a consistent way.
The container state add and removes ungrouped and grouped states with the right flags
and method calls to reuse the most of the old models.
The insert meta data method checks and gives feedback to reused models
according the handed logger method for inconsistencies (so for group_states an error
and for ungroup an info message).