fix(main window ctrl): destruction by explicit destroy console widget, too
The console was also not releasing it signal handler so it handled like the side bars.
Left a todo (pattern "TODO D-") at this place to investigate this further and not forget about it.
The debugging code is left in the test to re-use it for debugging the left over core element
in case all tests are running in a row.