fix(menu-bar/execution engine): secure active state machine id for "run until state"
The fix includes a securing mechanism on menu-bar gui site and the core.
On core site the order of setting active state machine id and the execution status was
logically wrong. On menu bar site the method was not checking if the sactive state machine id
is maybe None.
This can be seen as a work around but whithout refactoring active state machine id
usage at all it seems to be the best solution known issue #495. Its put into the develop
to avoid releasing this bug in the next patch.
Add changelog accordingly.