chore(gui): add logger message verbose and info level to track sm-opening time
The restore session feature also integrates duration logger messages to observe loading time.
As long as the restore has a wait for gui this duration includes the full opening and
the drawing time of e.g. gaphas.
Also to logger message of verbose level are add in the gaphas editor to show
method runtime duration of __init__ and setup_canvas methods.