feat(debug console): add first version of a follow mode
The follow mode is activatable and works but the disabling cause and swithces of
debug modes cause the scroller to move done even if it should not.
In case of the follow button checkbox press that should be avoided.

Note: This feature is part of the refactoring branch because it needs the refactoring and
should guide the refactoring. The feature results from issue #520.