feat(logging text view): keep cursor position after update
This allows to switch between follow and not follow mode and resumee on position
where the user was before. So to temporary follow execution log enable follow mode
and to come the previous read line disable follow mode and the view jumps back.

Note: A todo is left to keep track of the line if logger mode is changed.