feat(states editor): observes source editor cursor position and adapts pane
The implementation control's something (pane) of the main window controller
which is the parent controller and according changes of its child-child source editor
controller.view. As long as the States-Editor is always in the GUI it is assumed
this broke of encapsulation is proportionate to the difficulty of the problem of
issue #515.