refactor(source editor view): add generic button_container_min_size property
The property is summing up the button widths so that the allocation width is not depended
resize while runtime.
It can not be evaluated before runtime because of necessary use of get_allocation method for
external and pylint buttons.
Include it into the states editor method and make the minimal text width depended to it.