fix(debug_console test): add a work around to create warnings annd no assertion
The assertion was not closing the gui in a proper way.
The warning should secure it but let the test still fail.
Add wait for gui calls to hopefully solve the issue by that.
The gailure is only occuring on the build-bot for now so
I have to debug on the repo and push the changes.