fix(remove state machine): in state machines controller and state machine manager
Let the state machine manager return if called with id of state machine which is no more
in state machines dictionary.
And do not let the state machines controller call the remove operation a second time if
the state machine is removed by refresh and thereby was already removed with the helper function.

This changes should fix or improve issue #552.