fix(PortView): Define initial values for last_label_size
This fixes #571. Scoped variables are now properly drawn after loading.

The last_label_size variable is usually set during a drawing operation.
Its value is also used by a constraint to calculate the distance between
the port and the perpendicular waypoint. This constraint was solved
before the first drawing operation and thus used the wrong size of the
label.