fix(expected future models): notify a warning with additional information
The previous practice to assert in case of not empty expected future model lists
in cases in which it should be empty was not helping to solve the issues. This mostly was related
to no reports. Therefore a message is giving the user some insides and encourage them to
help resolving the issue by reporting the reproduction sheme.
The new helper function by default also cleans the respective elements
to avoid screwing the users work and thereby reduce RAFCONs usability.