feat(calculate closest sibling): add optional from logical port
The method can also do the calculation from a logicla port kind handed as
from_logical_port 'outcome' or 'income' of handed state_m.
In case of 'outcome' the first free port is used otherwise the first common one.
If the parameter is None the state is taken as reference position and size.