fix(state machine tree): restore expansion state tolerate missing states in libraries
The check for normal states in not generated library state state_copy models
and that those could be missing, too, was missing. It is now in general tolerated
in the restore of the expansion state that there could be missing states
if a library root state is above.

This is related to issue #538, too.