fix(source editor): right handling of successfull ext-editor opening
Also the failing of the command handed from the config file has changed.
If the command fails it get automatically reset. The failing command is shown in the debug console on debug level.
Now after a failing ext-editor opening, the user is automatically ask for a valid external editor.

This solve #596.