feat(group states): give the action a usefull selection
Generally the continues after group to adapt the new hierarchy state therefore it is selected.
Follow up actions are often resize if the new state or rename it (by e.g. <F2>).

The other option to keep the selection was not the case before and mostly would not be helpful
therefore it was changed.