refactor(state_machines_Editor): Remove PlusAddNotebook - fixes #639
The purpose of the PlusAddNotebook was to show a "+" button next to the
tabs for adding a new state machine. This feature was broken, because of
the way drawing works in Gtk3. The "on_draw" method was drawing behind
the widget, which is why the drawing (the +) was not visible.

We decided on putting not too much work onto this issue, as there are
multiple more ways to add a new state machine. This is why the custom
widget is now fully dropped.