refactor(complex action usage): add ComplexActionObserver and move data structure
The data structure is mostly the same but it is situated now in the state machine model.
The dictionary ongoing_complex_actions is an observable of the state machine model.
The ComplexActionObserver encapsulate the observation of this actions and
avoids to mix those with the root state observation of the state machine model.