feat(library model): ofers method recursive_model_generation
This method can be called multiple times and the models get extended to new depth.
This also can be used to have adaptive depth of libraries if gaphas can handle this.

This enables to remove the work around of that generates fully new state models and transfered the meta data.