feat(flat_storage): Support flat storage for single-state machines
In this common use case, where a library state-machine consists
of just one ExecutionState, the regular folder structure results in
a lot of duplicated information, and is nested an extra level unnecessarily.
These changes these kinds of state machines to be stored in a single directory
with no subdirectories. Storage code evaluates whether a state machine is of this type
by noticing the absence of "root_state_storage_id" and looking for a state in the current
folder in this case. Saving will adopt this format if the state machine contains exactly one
state and the state is an ExecutionState