Refactoring
Built-in function `len` used as condition
Using the len function to check if a sequence is empty is not idiomatic and can be less performant than checking the truthiness of the object.
len doesn't know the context in which it is called, so if computing the length means traversing the entire sequence, it must; it doesn't know that the result is just being compared to 0. Computing the boolean value can stop after it sees the first element, regardless of how long the sequence actually is.
Unnecessary use of `comprehension`
It is unnecessary to use a comprehension just to loop over the iterable and create a list/set/dict out of it. Python has a specialized set of tools for this task: the list/set/dict constructors, which are faster and more readable.

Use of `len(seq) - 1` to get last element of an iterable
There’s no need to calculate length of an iterable in order to fetch the last element of the iterable. You can provide a negative index -1 to it directly in orger to get the last element. In this way, you don't have to iterate over the sequence using len to get the last index when your purpose is only to get the last element.

`hasattr` used to check if the object is callable
It is unreliable to use hasattr(object, '__call__') to check if an object is callabe as it can give wrong results if object implements custom __getattr__ or its __call__ is itself not callabe. It is recommended to use the callable built-in instead.

Use of an unneeded `not` detected