feat(layout): Now shows in modification history
When performing auto-layouting, the current implementation would automatically save the state machine after the action and refresh the selected state. This is problematic as it erases the modification history, so any auto-layouting could not be undone once clicked. Now, the auto-layout action will show in the modification histroy and the state machine is not automatically saved.