fix(gui): Fixing segmentation fault when changing state type
The segmentation fault is caused when exiting the StateOverviewController because the GTK interfaces calls a destructor for the controller. However, the old controller is already destroyed during the changing of state type. This is also the reason why it works with the right-click option, because here we are not in the scope of the StateOverviewController and no (additional) destructor is called.
As a solution: When changing the state type over the state editor (menu on the right) the StateOverviewController is not destructed for the old state. This is definitely not the cleanest/most proper fix but the alternative would be to do some major restructuring in the StateOverviewController.