fix(gui): Fixed issue where CTRL+Z would undo twice
When pressing CTRL+Z in the source editor or when changing the description of a state, the undo action was triggered twice (same for redo). The problem is that the callback_function must return true to signalize the event was handled and does not need to be propagated further. (Source: https://bugs.webkit.org/show_bug.cgi?id=136430#c1)