fix(gui): Fixed bug with data ports outside parent
This fixes the issue where the data connection error, when trying to connect a data port output with a data port input outside of the parent state, was not handled properly and could potentially corrupt the state machine. Now, a meaningful error is raised and the connection is deleted properly.