feat(execution_history): Added locking current view feature
This feature adds a locking checkbox in the execution history viewer to enable to switch state machines in the graphical editor while keeping the execution history locked. This should make debugging a little bit easier and enable more flexible switching during debugging.
The other buttons "Open externall", "Clean" and "Reload" then work with respect to the locked execution history. So what is shown in the GUI in the locked state will be opened, cleaned and reloaded respectively.